DOC: Move the `README` contents back to the file and link from doc
Move the `README` contents back to the file and link from the `ìndex.rst`
documentation file through the `doc/readme_link.rst` file.

Add the documentation status badge to the `README` file.

Add the library instructions to the `About` section of the `README` file.

Make the `getting_started.rst` file line length be limited to 79 characters.